// College_App/src/theme/types.ts
export interface Room {
    id: number;
    room_number: string;
    type: string;
    capacity: number;
    status: 'Available' | 'Booked' | 'Pending';
}

export interface User {
    id: number;
    name: string;
    email: string;
    role: 'student' | 'faculty' | 'admin';
}

export interface Booking {
    id: number;
    booking_date: string;
    time_slot: string;
    status: 'Confirmed' | 'Cancelled' | 'Pending';
    user_id: number;
    room_id: number;
}